Master now has #351, which added ShutdownDuringDownload as donedl case 9.
The queue-budget cases move to 10 (map change) and 11 (donedl), and the
file header and the runner's case loop list both issues and cases 0-11.

+++ code/client/snd_mem.c
@@ -128,7 +128,7 @@ static int GetLittleLong(void)
 	val = *data_p;
 	val = val + (*(data_p+1)<<8);
 	val = val + (*(data_p+2)<<16);
-	val = val + (*(data_p+3)<<24);
+	val = val + (int)((unsigned int)*(data_p+3)<<24);
 	data_p += 4;
 	return val;
 }
